Added endpoint for application, and started implementing endpoints + services + exceptions + rest advice for creating and viewing all existing availability periods and competence profiles for a specific person, changed competence profile yearsOfExperience data type from BigDecimal to double since that data type has a parser while BigDecimal does not

+++ SpringProgram/demo/src/main/java/com/example/demo/domain/dto/CompetenceProfileDTO.java
@@ -1,8 +1,5 @@
 package com.example.demo.domain.dto;
 
-import java.math.BigDecimal;
-
-
 /** 
  * The CompetenceProfileDTO interface provides a data transfer interface for handling competence profile objects
 */
@@ -30,5 +27,5 @@ public interface CompetenceProfileDTO {
      * 
      * @return the number of years of experience
      */
-    public BigDecimal getYearsOfExperience();
+    public double getYearsOfExperience();
 }
